# lib/api-zod/src/generated/types/protocolStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arena Protocol API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProtocolStats {
  /** Distinct registered players */
  activeFighters: number;
  /** Total battles recorded */
  totalBattles: number;
  /** Total ARENA paid out across all wins, as a decimal string */
  totalRewardsArena: string;
  /** Active marketplace listings */
  activeListings: number;
}